Uptime Debug
Corrected runtime error by first converting the call to time.time() to an integer for the purpose of dropping the unnescessary decimals. Then rewriting the javascript to properly post and format the calculated runtime. (Some operation added to javascript to solve a decimal issue that occured in the javascript code.